+++ sketch-extensions/src/main/java/com/github/panpf/sketch/drawable/MaskProgressDrawable.kt
@@ -15,11 +15,11 @@
  */
 package com.github.panpf.sketch.drawable
 
-import android.animation.ValueAnimator
 import android.graphics.Canvas
 import android.graphics.ColorFilter
 import android.graphics.Paint
 import android.graphics.PixelFormat
+import android.os.SystemClock
 import androidx.annotation.ColorInt
 import com.github.panpf.sketch.util.format
 
@@ -28,10 +28,12 @@ import com.github.panpf.sketch.util.format
  */
 class MaskProgressDrawable(
     @ColorInt private val maskColor: Int = DEFAULT_MASK_COLOR
+    // todo 增加等待时显示控制参数
 ) : ProgressDrawable() {
 
     companion object {
         const val DEFAULT_MASK_COLOR = 0x22000000
+        private const val DEFAULT_DURATION = 300
     }
 
     private val paint = Paint().apply {
@@ -39,49 +41,71 @@ class MaskProgressDrawable(
         isAntiAlias = true
     }
 
+    private var animationRunning: Boolean = false
+    private var animationStartProgress: Float? = null
+    private var animationTargetProgress: Float? = null
+    private var animationStartTimeMillis = 0L
     private var _progress: Float = 0f
         set(value) {
             field = value
-            invalidateSelf()
             if (value >= 1f) {
                 onProgressEnd?.invoke()
             }
         }
-
-    private var progressAnimator: ValueAnimator? = null
-
     override var progress: Float
         get() = _progress
         set(value) {
-            val newValue = value.format(1).coerceAtLeast(0f).coerceAtMost(1f)
+            val newValue = value.format(1).coerceIn(0f, 1f)
             if (newValue != _progress) {
                 if (_progress == 0f && newValue == 1f) {
                     // Here is the loading of the local image, no loading progress, quickly complete
                     _progress = newValue
+                    animationRunning = false
                 } else if (newValue > _progress) {
-                    animationUpdateProgress(newValue)
+                    animationStartProgress = _progress
+                    animationTargetProgress = newValue
+                    animationStartTimeMillis = SystemClock.uptimeMillis()
+                    animationRunning = true
                 } else {
                     // If newValue is less than _progress, you can reset it quickly
                     _progress = newValue
+                    animationRunning = false
                 }
+                invalidateSelf()
             }
         }
+
     override var onProgressEnd: (() -> Unit)? = null
 
     override fun draw(canvas: Canvas) {
-        val currentProgress = _progress.takeIf { it > 0f } ?: return
+        var animationDone = false
+        if (animationRunning) {
+            val percent =
+                (SystemClock.uptimeMillis() - animationStartTimeMillis) / DEFAULT_DURATION.toDouble()
+            animationDone = percent >= 1
+            _progress =
+                (animationStartProgress!! + ((animationTargetProgress!! - animationStartProgress!!) * percent)).toFloat()
+        }
+
+        val progress = _progress.takeIf { it > 0f } ?: return
         val bounds = bounds.takeIf { !it.isEmpty } ?: return
         val saveCount = canvas.save()
-
         canvas.drawRect(
             bounds.left.toFloat(),
-            bounds.top + ((currentProgress * bounds.height())),
+            bounds.top + ((progress * bounds.height())),
             bounds.right.toFloat(),
             bounds.bottom.toFloat(),
             paint
         )
-
         canvas.restoreToCount(saveCount)
+
+        if (animationRunning) {
+            if (animationDone) {
+                animationRunning = false
+            } else {
+                invalidateSelf()
+            }
+        }
     }
 
     override fun setAlpha(alpha: Int) {
@@ -92,27 +116,17 @@ class MaskProgressDrawable(
         paint.colorFilter = colorFilter
     }
 
+    @Deprecated(
+        "Deprecated in Java. This method is no longer used in graphics optimizations",
+        ReplaceWith("")
+    )
     override fun getOpacity(): Int = PixelFormat.TRANSLUCENT
-
-    private fun animationUpdateProgress(newProgress: Float) {
-        progressAnimator?.cancel()
-        progressAnimator = ValueAnimator.ofFloat(_progress, newProgress).apply {
-            addUpdateListener {
-                if (isActive()) {
-                    _progress = animatedValue as Float
-                } else {
-                    it?.cancel()
-                }
-            }
-            duration = 300
-        }
-        progressAnimator?.start()
-    }
 
     override fun setVisible(visible: Boolean, restart: Boolean): Boolean {
         val changed = super.setVisible(visible, restart)
         if (changed && !visible) {
-            progressAnimator?.cancel()
+            animationRunning = false
+            invalidateSelf()
         }
         return changed
     }