+++ sketch-extensions/src/main/java/com/github/panpf/sketch/drawable/SectorProgressDrawable.kt
@@ -15,7 +15,6 @@
  */
 package com.github.panpf.sketch.drawable
 
-import android.animation.ValueAnimator
 import android.content.res.Resources
 import android.graphics.Canvas
 import android.graphics.Color
@@ -25,6 +24,7 @@ import android.graphics.Paint.Style.FILL
 import android.graphics.Paint.Style.STROKE
 import android.graphics.PixelFormat
 import android.graphics.RectF
+import android.os.SystemClock
 import com.github.panpf.sketch.util.format
 
 /**
@@ -37,6 +37,10 @@ class SectorProgressDrawable(
     private val progressColor: Int = Color.WHITE,
     private val strokeWidth: Float = size * 0.02f,
 ) : ProgressDrawable() {
+
+    companion object {
+        private const val DEFAULT_DURATION = 300
+    }
 
     private val backgroundPaint = Paint().apply {
         isAntiAlias = true
@@ -55,51 +59,52 @@ class SectorProgressDrawable(
     }
     private val progressOval = RectF()
 
+    private var animationRunning: Boolean = false
+    private var animationStartProgress: Float? = null
+    private var animationTargetProgress: Float? = null
+    private var animationStartTimeMillis = 0L
     private var _progress: Float = 0f
         set(value) {
             field = value
-            invalidateSelf()
             if (value >= 1f) {
                 onProgressEnd?.invoke()
             }
         }
-
-    private var progressAnimator: ValueAnimator? = null
-
     override var progress: Float
         get() = _progress
         set(value) {
-            val newValue = value.format(1).coerceAtLeast(0f).coerceAtMost(1f)
+            val newValue = value.format(1).coerceIn(0f, 1f)
             if (newValue != _progress) {
                 if (_progress == 0f && newValue == 1f) {
                     // Here is the loading of the local image, no loading progress, quickly complete
                     _progress = newValue
+                    animationRunning = false
                 } else if (newValue > _progress) {
-                    animationUpdateProgress(newValue)
+                    animationStartProgress = _progress
+                    animationTargetProgress = newValue
+                    animationStartTimeMillis = SystemClock.uptimeMillis()
+                    animationRunning = true
                 } else {
                     // If newValue is less than _progress, you can reset it quickly
                     _progress = newValue
+                    animationRunning = false
                 }
+                invalidateSelf()
             }
         }
+
     override var onProgressEnd: (() -> Unit)? = null
 
-    private fun animationUpdateProgress(newProgress: Float) {
-        progressAnimator?.cancel()
-        progressAnimator = ValueAnimator.ofFloat(_progress, newProgress).apply {
-            addUpdateListener {
-                if (isActive()) {
-                    _progress = animatedValue as Float
-                } else {
-                    it?.cancel()
-                }
-            }
-            duration = 300
-        }
-        progressAnimator?.start()
-    }
-
     override fun draw(canvas: Canvas) {
+        var animationDone = false
+        if (animationRunning) {
+            val percent =
+                (SystemClock.uptimeMillis() - animationStartTimeMillis) / DEFAULT_DURATION.toDouble()
+            animationDone = percent >= 1
+            _progress =
+                (animationStartProgress!! + ((animationTargetProgress!! - animationStartProgress!!) * percent)).toFloat()
+        }
+
         val progress = _progress.takeIf { it >= 0f } ?: return
         val bounds = bounds.takeIf { !it.isEmpty } ?: return
         val saveCount = canvas.save()
@@ -127,6 +132,14 @@ class SectorProgressDrawable(
         canvas.drawArc(progressOval, 270f, sweepAngle, true, progressPaint)
 
         canvas.restoreToCount(saveCount)
+
+        if (animationRunning) {
+            if (animationDone) {
+                animationRunning = false
+            } else {
+                invalidateSelf()
+            }
+        }
     }
 
     override fun setAlpha(alpha: Int) {
@@ -141,12 +154,17 @@ class SectorProgressDrawable(
         progressPaint.colorFilter = colorFilter
     }
 
+    @Deprecated(
+        "Deprecated in Java. This method is no longer used in graphics optimizations",
+        ReplaceWith("")
+    )
     override fun getOpacity(): Int = PixelFormat.TRANSLUCENT
 
     override fun setVisible(visible: Boolean, restart: Boolean): Boolean {
         val changed = super.setVisible(visible, restart)
         if (changed && !visible) {
-            progressAnimator?.cancel()
+            animationRunning = false
+            invalidateSelf()
         }
         return changed
     }